Added elliptical orbits
Added the option to change the eccentricity of the orbit. 0 < ecc < 0.25 for LEO, with ecc = 0 being a circular orbit. Using variable dist2CoE (distance to center of Earth) as a focus of the ellipse and using a given eccentricity to calculate the semi major axis, and then the initial velocity (v0) of the satellite.